==> Capstone_Two/1. Data Access.ipynb <==
# Data Access

This notebook served to explore the Twitter API using Tweepy and pull timeline tweets from 10 selected Twitter accounts. At the end of the notebook the data is saved in a CSV file for use in subsequent notebooks.

#access keys removed after running

import os
import tweepy as tw
import pandas as pd

consumer_key = ''
consumer_key_secret = ''
access_token= ''
access_token_secret = ''

Bearer_token = ''

auth = tw.OAuthHandler(consumer_key, consumer_key_secret)
auth.set_access_token(access_token, access_token_secret)

api = tw.API(auth, wait_on_rate_limit=True)

#users selected by referencing https://www.music-map.com/ and spotify to determine related artists
#minimum 5k tweets per account and all verified accounts

screen_names = ['awonderland','wearegalantis','sanholobeats','theknocks','pekingduk','porterrobinson','WhatSoNot','LouisTheChild','NGHTMRE','JaiWolfx']

#https://gist.github.com/yanofsky/5436496 referenced for get_timeline_tweets function
#https://gist.github.com/jaymcgrath/367c521f1dd786bc5a05ec3eeeb1cb04 used to view list of tweepy status object attributes

def get_timeline_tweets(screen_name):
    tweets = []
    new_tweets = api.user_timeline(screen_name = screen_name,count=200)
    tweets.extend(new_tweets)
    last_id = new_tweets[-1].id - 1
    
    while len(new_tweets) > 0:
        new_tweets = api.user_timeline(screen_name = screen_name,count=200,max_id = last_id)
        tweets.extend(new_tweets)
        
        if len(new_tweets) > 0:
            last_id = new_tweets[-1].id - 1
    
    tweet_data = [[tweet.user.screen_name, tweet.in_reply_to_screen_name, tweet.is_quote_status, tweet.text,tweet.created_at,
                   tweet.favorite_count,tweet.retweet_count, tweet.author.followers_count,tweet.entities,tweet.truncated,
                   tweet.author.statuses_count] for tweet in tweets]
    
    columns = ['screen_name','reply_to','is_quote_status','text','created_at','favorites','retweets','follower_count',
               'entities','is_truncated','statuses_count']
    
    return pd.DataFrame(data=tweet_data,columns = columns)

tweet_data = pd.DataFrame()

for screen_name in screen_names:
    tweet_data = tweet_data.append(get_timeline_tweets(screen_name))

tweet_data.describe()

tweet_data.head()

tweet_data.info()

Exploring the is_truncated attribute.

tweet_data[tweet_data.is_truncated==True].head()

tweet_data.iloc[243].text

tweet_data.iloc[240].entities['hashtags']

tweet_data.to_csv('./tweet_data_Feb1.csv')
