==> Capstone_Two/4. NLP Data Processing.ipynb <==
# NLP Data Processing

This notebook serves to perform beginner NLP pre-processing by using a simple CountVectorizer on the tweet texts and adding the word counts as features to the ultimately exported csv file to be used for model training.

import pandas as pd
import nltk
from nltk import sent_tokenize
from nltk import word_tokenize
#nltk.download('punkt')
#nltk.download('stopwords')
from nltk.corpus import stopwords

stopwords = stopwords.words('english')

data = pd.read_csv('clean_tweet_data.csv',index_col=0,parse_dates=['created_at'])

data.head()

Many of the tweets contain an http.* string that references either an image or a video. I will replace all of these with a custom containsmedia string to include as another feature for the model.

data['text'] = data.text.str.replace('http.*','containsmedia',regex=True)
data['time_of_day'] = data['created_at'].dt.hour

data.head()

Create a DataFrame with a count of the words in each tweet:

word_list = pd.DataFrame()

for i in range(len(data)):
    clean_words = []
    temp_df = pd.DataFrame()
    for word in word_tokenize(data.text[i]):
        if word not in stopwords and word.isalpha():
            clean_words.append(word.lower())
            faves = data.favorites[i]
            rts = data.retweets[i]
            temp_df = pd.DataFrame(data = {'Word':clean_words,'Favorites':faves,'Retweets':rts})
    word_list = word_list.append(temp_df)

tweet_words = word_list.groupby(['Word']).sum().sort_values(by='Favorites',ascending=False)

word_counts = word_list.groupby(['Word']).count().sort_values(by='Favorites',ascending=False)

final_word_data = tweet_words.join(word_counts,on='Word',rsuffix=' tweet count')

Looking at some of the most commonly used words and their associated tweets performances:

final_word_data

final_word_data['Rts/Tweet'] = final_word_data.Retweets/final_word_data['Retweets tweet count']
final_word_data['Faves/Tweet'] = final_word_data.Favorites/final_word_data['Favorites tweet count']

final_word_data[final_word_data['Favorites tweet count'] > 100].sort_values('Faves/Tweet',ascending = False).head(20)

final_word_data.index.values

Total number of words:

len(final_word_data.index.values)

I will need to reduce the number of features so I will only look at words used at least 100 times, which gives a much more reasonable 115 vectors.

len(final_word_data[final_word_data['Favorites tweet count']>100].index.values)

Creating a DataFrame using only the words used over 100 times and not including stopwords:

from sklearn.feature_extraction.text import CountVectorizer

vocab = final_word_data[final_word_data['Favorites tweet count']>100].index.values
vocab = [word for word in vocab if not word in stopwords]

vectorizer = CountVectorizer(vocabulary = vocab)

X = []

for i in range(0,len(data)):
    X.append(vectorizer.fit_transform([data.text[i]]))

X_arrays = []
for i in range(0,len(X)):
    X_arrays.append(X[i].toarray())

import numpy as np

feature_words = pd.DataFrame(np.concatenate(X_arrays),columns=vocab)

Inspecting the new feature words DataFrame:

feature_words.head()

Join the feature words DataFrame to the original data and export it as a csv for further pre-processing in the next notebook:

data_feat = data.join(feature_words,rsuffix = 'feat')

data_feat.head()

data_feat.to_csv('./data_wordvectorized.csv')
