==> Capstone_Two/5. Pre-processing.ipynb <==
# Pre-processing

This notebook serves to process all the features so that they can be easily used to build models in a subsequent notebook. The result of this notebook is 4 csv files containing X and y data split into training and test sets.

import pandas as pd
from sklearn.model_selection import train_test_split

data = pd.read_csv('data_wordvectorized.csv',index_col=0,parse_dates=['created_at'])

data.head()

Remove unecessary features.

model_data = data.drop(columns=['is_quote_status','text','created_at','follower_count','is_truncated','statuses_count'])
model_data.head()

Get dummy variables for the categorical variables and drop the first to avoid collinearity:

model_data = pd.get_dummies(model_data,columns=['screen_name','Day_of_Week','time_of_day'],drop_first=True)

List of features currently:

model_data.columns.values

Manually select most relevant features for the model:

x_columns = ['containsmedia', 'music', 'new',
       'love', 'today', 'first', 'like', 'live', 'song', 'good', 'time', 'show', 'day',
       'tomorrow', 'set', 'virtual', 'people', 'miss', 'album', 'going', 'tonight', 'happy', 'guys',
       'ever', 'year', 'tour', 'play', 'life', 'last', 'best', 'night', 'video', 
        'screen_name_LouisTheChild', 'screen_name_NGHTMRE', 'screen_name_WhatSoNot', 'screen_name_awonderland',
       'screen_name_pekingduk', 'screen_name_porterrobinson', 'screen_name_sanholobeats', 'screen_name_theknocks',
       'screen_name_wearegalantis', 'Day_of_Week_1', 'Day_of_Week_2', 'Day_of_Week_3', 'Day_of_Week_4', 'Day_of_Week_5', 
        'Day_of_Week_6', 'time_of_day_1', 'time_of_day_2', 'time_of_day_3', 'time_of_day_4', 'time_of_day_5', 'time_of_day_6', 
        'time_of_day_7', 'time_of_day_8', 'time_of_day_9', 'time_of_day_10', 'time_of_day_11',
       'time_of_day_12', 'time_of_day_13', 'time_of_day_14',
       'time_of_day_15', 'time_of_day_16', 'time_of_day_17',
       'time_of_day_18', 'time_of_day_19', 'time_of_day_20',
       'time_of_day_21', 'time_of_day_22', 'time_of_day_23']

len(x_columns)

70 features is much more manageable.

model_data.head()

X = model_data[x_columns]

y = model_data[['favorites','retweets']]

Split data into test (25%) and training (75%) sets and save as csv files:

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=42)

X_train.to_csv('.\X_train_v2.csv')
X_test.to_csv('.\X_test_v2.csv')
y_train.to_csv('.\y_train_v2.csv')
y_test.to_csv('.\y_test_v2.csv')
