==> Capstone_Two/7. Modeling 2.ipynb <==
# Modeling 2

This notebook investigates if separating the data by artist leads to improved predictive capability.

import pandas as pd
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_squared_error
from sklearn.model_selection import train_test_split
import time

X_train = pd.read_csv('Data/X_train_v2.csv',index_col = 0)
X_test = pd.read_csv('Data/X_test_v2.csv',index_col = 0)
y_train = pd.read_csv('Data/y_train_v2.csv',index_col = 0)
y_test = pd.read_csv('Data/y_test_v2.csv',index_col = 0)

X = pd.concat([X_train,X_test])
y = pd.concat([y_train,y_test])

## Build Linear Models for Each Artist

def LinearRegressor(X_train,X_test,y_train,y_test):
    lr = LinearRegression()

    lr_fit_start = time.time()
    lr.fit(X_train,y_train['favorites'])
    lr_fit_end = time.time()
    lr_fit_time = lr_fit_end - lr_fit_start

    lr_pred_start = time.time()
    lr_pred = lr.predict(X_test)
    lr_pred_end = time.time()
    lr_pred_time = lr_pred_end - lr_pred_start

    mse_lr_test = mean_squared_error(y_test['favorites'],lr_pred)
    mse_lr_train = mean_squared_error(y_train['favorites'],lr.predict(X_train))

    print('The train mse for the linear regression model is ' + str(round(mse_lr_train,3)))
    print('The test mse for the linear regression model is ' + str(round(mse_lr_test,3)))
    print('The model fitting time is ' + str(round(lr_fit_time,3)))
    print('The model prediction time is ' + str(round(lr_pred_time,3)))
    
    return mse_lr_train, mse_lr_test

alison_x =  X[X.screen_name_awonderland==1]
alison_y = y.loc[alison_x.index]
X_train, X_test, y_train, y_test = train_test_split(alison_x, alison_y, test_size=0.25, random_state=42)
alison_train_mse, alison_test_mse = LinearRegressor(X_train, X_test, y_train, y_test)

ltc_x =  X[X.screen_name_LouisTheChild==1]
ltc_y = y.loc[ltc_x.index]
X_train, X_test, y_train, y_test = train_test_split(ltc_x, ltc_y, test_size=0.25, random_state=42)
ltc_train_mse, ltc_test_mse = LinearRegressor(X_train, X_test, y_train, y_test)

nght_x =  X[X.screen_name_NGHTMRE==1]
nght_y = y.loc[nght_x.index]
X_train, X_test, y_train, y_test = train_test_split(nght_x, nght_y, test_size=0.25, random_state=42)
nght_train_mse, nght_test_mse = LinearRegressor(X_train, X_test, y_train, y_test)

what_x =  X[X.screen_name_WhatSoNot==1]
what_y = y.loc[what_x.index]
X_train, X_test, y_train, y_test = train_test_split(what_x, what_y, test_size=0.25, random_state=42)
what_train_mse, what_test_mse = LinearRegressor(X_train, X_test, y_train, y_test)

duk_x =  X[X.screen_name_pekingduk==1]
duk_y = y.loc[duk_x.index]
X_train, X_test, y_train, y_test = train_test_split(duk_x, duk_y, test_size=0.25, random_state=42)
duk_train_mse, duk_test_mse = LinearRegressor(X_train, X_test, y_train, y_test)

porter_x =  X[X.screen_name_porterrobinson==1]
porter_y = y.loc[porter_x.index]
X_train, X_test, y_train, y_test = train_test_split(porter_x, porter_y, test_size=0.25, random_state=42)
porter_train_mse, porter_test_mse = LinearRegressor(X_train, X_test, y_train, y_test)

san_x =  X[X.screen_name_sanholobeats==1]
san_y = y.loc[san_x.index]
X_train, X_test, y_train, y_test = train_test_split(san_x, san_y, test_size=0.25, random_state=42)
san_train_mse, san_test_mse = LinearRegressor(X_train, X_test, y_train, y_test)

knocks_x =  X[X.screen_name_theknocks==1]
knocks_y = y.loc[knocks_x.index]
X_train, X_test, y_train, y_test = train_test_split(knocks_x, knocks_y, test_size=0.25, random_state=42)
knocks_train_mse, knocks_test_mse = LinearRegressor(X_train, X_test, y_train, y_test)

galantis_x =  X[X.screen_name_wearegalantis==1]
galantis_y = y.loc[galantis_x.index]
X_train, X_test, y_train, y_test = train_test_split(galantis_x, galantis_y, test_size=0.25, random_state=42)
galantis_train_mse, galantis_test_mse = LinearRegressor(X_train, X_test, y_train, y_test)

jai_x =  X[(X.screen_name_awonderland==0) & (X.screen_name_wearegalantis==0) & (X.screen_name_theknocks==0) & (X.screen_name_sanholobeats==0) & (X.screen_name_porterrobinson==0) & (X.screen_name_pekingduk==0) & (X.screen_name_WhatSoNot==0) & (X.screen_name_NGHTMRE==0) & (X.screen_name_LouisTheChild==0)]
jai_y = y.loc[jai_x.index]
X_train, X_test, y_train, y_test = train_test_split(jai_x, jai_y, test_size=0.25, random_state=42)
jai_train_mse, jai_test_mse = LinearRegressor(X_train, X_test, y_train, y_test)

artists = ['awonderland','wearegalantis','sanholobeats','theknocks','pekingduk','porterrobinson','WhatSoNot','LouisTheChild','NGHTMRE','JaiWolfx']
train_mse = [alison_train_mse,galantis_train_mse,san_train_mse,knocks_train_mse,duk_train_mse,porter_train_mse,what_train_mse,ltc_train_mse,nght_train_mse,jai_train_mse]
test_mse = [alison_test_mse,galantis_test_mse,san_test_mse,knocks_test_mse,duk_test_mse,porter_test_mse,what_test_mse,ltc_test_mse,nght_test_mse,jai_test_mse]
train_size = [len(alison_x),len(galantis_x),len(san_x),len(knocks_x),len(duk_x),len(porter_x),len(what_x),len(ltc_x),len(nght_x),len(jai_x)]


results = pd.DataFrame({'Artist':artists,'Training Size':train_size,'Train MSE':train_mse,'Test MSE':test_mse})

## Results

results.set_index('Artist',inplace=True)
results = results.astype(int)
results

Some artists have much more predictable favorites while others are quite unpredictable. Further investigation would have to be done to determine the cause. One hypothesis I have is that the lower error artists simply have fewer viral tweets, which are much harder for a model to predict.
